refactor: 테스트 코드 package 이동
