+++ src/main/java/chess/chessboard/position/Direction.java
@@ -1,8 +1,8 @@
-package chess.chessBoard;
+package chess.chessboard.position;
 
-import chess.chessBoard.position.File;
-import chess.chessBoard.position.Position;
-import chess.chessBoard.position.Rank;
+import chess.chessboard.position.File;
+import chess.chessboard.position.Position;
+import chess.chessboard.position.Rank;
 
 public enum Direction {
