+++ src/main/java/chess/chessboard/position/Position.java
@@ -1,6 +1,4 @@
-package chess.chessBoard.position;
-
-import chess.chessBoard.Direction;
+package chess.chessboard.position;
 
 import java.util.Objects;
