+++ src/main/java/chess/piece/Pawn.java
@@ -1,13 +1,13 @@
 package chess.piece;
 
-import chess.chessBoard.Direction;
+import chess.chessboard.position.Direction;
 import chess.game.Player;
-import chess.chessBoard.position.Position;
+import chess.chessboard.position.Position;
 
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static chess.chessBoard.Direction.*;
+import static chess.chessboard.position.Direction.*;
 import static chess.game.Player.BLACK;
 import static chess.game.Player.NONE;
