+++ src/main/java/chess/piece/Queen.java
@@ -1,13 +1,13 @@
 package chess.piece;
 
-import chess.chessBoard.Direction;
+import chess.chessboard.position.Direction;
 import chess.game.Player;
-import chess.chessBoard.position.Position;
+import chess.chessboard.position.Position;
 
 import java.util.*;
 
-import static chess.chessBoard.Direction.*;
-import static chess.chessBoard.Direction.NORTH;
+import static chess.chessboard.position.Direction.*;
+import static chess.chessboard.position.Direction.NORTH;
 
 public final class Queen extends Piece {
