+++ src/main/java/chess/state/End.java
@@ -1,6 +1,6 @@
 package chess.state;
 
-import chess.chessBoard.Board;
+import chess.chessboard.Board;
 
 import static chess.state.Command.*;
