+++ src/main/java/chess/state/Running.java
@@ -1,6 +1,6 @@
 package chess.state;
 
-import chess.chessBoard.Board;
+import chess.chessboard.Board;
 
 public abstract class Running extends Started {
