+++ src/main/java/chess/state/Started.java
@@ -1,7 +1,7 @@
 package chess.state;
 
-import chess.chessBoard.Board;
-import chess.chessBoard.position.Position;
+import chess.chessboard.Board;
+import chess.chessboard.position.Position;
 import chess.piece.Piece;
 
 import java.util.Map;
