+++ src/main/java/chess/state/Status.java
@@ -1,6 +1,6 @@
 package chess.state;
 
-import chess.chessBoard.Board;
+import chess.chessboard.Board;
 import chess.game.Player;
 import chess.game.Score;
