+++ src/main/java/chess/state/Turn.java
@@ -1,8 +1,8 @@
 package chess.state;
 
-import chess.chessBoard.Board;
+import chess.chessboard.Board;
 import chess.game.Player;
-import chess.chessBoard.position.Position;
+import chess.chessboard.position.Position;
 
 import java.util.List;
