+++ src/test/java/chess/chessboard/BoardTest.java
@@ -1,9 +1,9 @@
-package chess;
+package chess.chessboard;
 
-import chess.chessBoard.Board;
-import chess.chessBoard.position.File;
-import chess.chessBoard.position.Rank;
-import chess.chessBoard.position.Position;
+import chess.chessboard.Board;
+import chess.chessboard.position.File;
+import chess.chessboard.position.Rank;
+import chess.chessboard.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
