+++ src/test/java/chess/piece/KingTest.java
@@ -1,10 +1,10 @@
 package chess.piece;
 
-import chess.chessBoard.Board;
-import chess.chessBoard.position.File;
-import chess.chessBoard.position.Rank;
+import chess.chessboard.Board;
+import chess.chessboard.position.File;
+import chess.chessboard.position.Rank;
 import chess.game.Player;
-import chess.chessBoard.position.Position;
+import chess.chessboard.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
